create FileSaverComponent
