+++ components/ActivityComponent.php
@@ -29,9 +29,11 @@ class ActivityComponent extends Component
     public function addActivity(Activity $activity) :bool
     {
         $activity->files = UploadedFile::getInstances($activity, 'files');
+        $fileSaver = \Yii::createObject(['class' => FileSaverComponent::class]);
         if ($activity->validate()) {
             foreach ($activity->files as &$file) {
-                $file = $this->saveFile($file);
+                $file = $fileSaver->saveFile($file);
+//                $file = $this->saveFile($file);
                 if (!$file) {
                     return false;
                 }
@@ -41,26 +43,6 @@ class ActivityComponent extends Component
         return false;
     }
 
-    private function saveFile(UploadedFile $file): ?string
-    {
-        $name = $this->genFileName($file);
-        $path = $this->getPathToSave() . $name;
-        if ($file->saveAs($path)) {
-            return $name;
-        }
-        return null;
-    }
 
-    private function getPathToSave()
-    {
-        $path = \Yii::getAlias('@webroot/files/');
-        FileHelper::createDirectory($path);
-        return  $path;
-    }
-
-    private function genFileName(UploadedFile $file)
-    {
-        return time() . '-' . $file->getBaseName() . '.' . $file->getExtension();
-    }
 
 }
